insights_functions/requirements.txt
google-cloud
google-cloud-vision
google-cloud-firestore
google-cloud-speech
google-cloud-language
grpcio